bq - refactored angle and distance calculation
